+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/BusBreakerVoltageLevel.java
@@ -308,7 +308,7 @@ class BusBreakerVoltageLevel extends AbstractVoltageLevel {
     protected final StateArray<StateImpl> states;
 
     BusBreakerVoltageLevel(String id, String name, SubstationImpl substation,
-                           float nominalV, float lowVoltageLimit, float highVoltageLimit) {
+                           double nominalV, double lowVoltageLimit, double highVoltageLimit) {
         super(id, name, substation, nominalV, lowVoltageLimit, highVoltageLimit);
         states = new StateArray<>(substation.getNetwork().getRef(), StateImpl::new);
         // invalidate topology and connected components
