+++ iidm/iidm-impl/src/test/java/com/powsybl/iidm/network/impl/CurrentLimitsTest.java
@@ -23,7 +23,7 @@ public class CurrentLimitsTest {
                 .add();
         VoltageLevel vl1 = s1.newVoltageLevel()
                 .setId("VL1")
-                .setNominalV(400f)
+                .setNominalV(400.0)
                 .setTopologyKind(TopologyKind.BUS_BREAKER)
                 .add();
         vl1.getBusBreakerView().newBus()
@@ -35,7 +35,7 @@ public class CurrentLimitsTest {
                 .add();
         VoltageLevel vl2 = s2.newVoltageLevel()
                 .setId("VL2")
-                .setNominalV(400f)
+                .setNominalV(400.0)
                 .setTopologyKind(TopologyKind.BUS_BREAKER)
                 .add();
         vl2.getBusBreakerView().newBus()
@@ -49,29 +49,29 @@ public class CurrentLimitsTest {
                 .setVoltageLevel2("VL2")
                 .setConnectableBus2("B2")
                 .setBus2("B2")
-                .setR(1)
-                .setX(1)
-                .setG1(0)
-                .setG2(0)
-                .setB1(0)
-                .setB2(0)
+                .setR(1.0)
+                .setX(1.0)
+                .setG1(0.0)
+                .setG2(0.0)
+                .setB1(0.0)
+                .setB2(0.0)
                 .add();
         l.newCurrentLimits1()
-                .setPermanentLimit(1000)
+                .setPermanentLimit(1000.0)
                 .beginTemporaryLimit()
                     .setName("20'")
                     .setAcceptableDuration(20 * 60)
-                    .setValue(1200)
+                    .setValue(1200.0)
                 .endTemporaryLimit()
                 .beginTemporaryLimit()
                     .setName("5'")
                     .setAcceptableDuration(5 * 60)
-                    .setValue(1400)
+                    .setValue(1400.0)
                 .endTemporaryLimit()
                 .beginTemporaryLimit()
                     .setName("1'")
                     .setAcceptableDuration(60)
-                    .setValue(1600)
+                    .setValue(1600.0)
                 .endTemporaryLimit()
             .add();
         return network;
@@ -82,34 +82,34 @@ public class CurrentLimitsTest {
         Network network = createNetwork();
         Line l = network.getLine("L");
         assertFalse(l.isOverloaded());
-        l.getTerminal1().getBusBreakerView().getBus().setV(390);
-        l.getTerminal1().setP(100).setQ(50); // i = 165.51212
-        assertFalse(Float.isNaN(l.getTerminal1().getI()));
+        l.getTerminal1().getBusBreakerView().getBus().setV(390.0);
+        l.getTerminal1().setP(100.0).setQ(50.0); // i = 165.51212
+        assertFalse(Double.isNaN(l.getTerminal1().getI()));
         assertFalse(l.isOverloaded());
         assertFalse(l.checkPermanentLimit1());
         assertNull(l.checkTemporaryLimits1());
 
-        l.getTerminal1().setP(800).setQ(400); // i = 1324.0969
+        l.getTerminal1().setP(800.0).setQ(400.0); // i = 1324.0969
         assertTrue(l.isOverloaded());
         assertEquals(5 * 60, l.getOverloadDuration());
         assertTrue(l.checkPermanentLimit1());
         assertNotNull(l.checkTemporaryLimits1());
         assertEquals(5 * 60, l.checkTemporaryLimits1().getTemporaryLimit().getAcceptableDuration());
-        assertEquals(1200.0f, l.checkTemporaryLimits1().getPreviousLimit(), 0.0f);
+        assertEquals(1200.0, l.checkTemporaryLimits1().getPreviousLimit(), 0.0);
 
-        l.getTerminal1().setP(900).setQ(500); // i = 1524.1499
+        l.getTerminal1().setP(900.0).setQ(500.0); // i = 1524.1499
         assertEquals(60, l.getOverloadDuration());
         assertNotNull(l.checkTemporaryLimits1());
         assertEquals(60, l.checkTemporaryLimits1().getTemporaryLimit().getAcceptableDuration());
-        assertEquals(1400.0f, l.checkTemporaryLimits1().getPreviousLimit(), 0.0f);
+        assertEquals(1400.0, l.checkTemporaryLimits1().getPreviousLimit(), 0.0);
     }
 
     @Test
     public void testLimitWithoutTempLimit() {
         Line l = createNetwork().getLine("L");
-        l.newCurrentLimits1().setPermanentLimit(1000.0f).add();
-        l.getTerminal1().getBusBreakerView().getBus().setV(390);
-        l.getTerminal1().setP(800).setQ(400); // i = 1324.0969
+        l.newCurrentLimits1().setPermanentLimit(1000.0).add();
+        l.getTerminal1().getBusBreakerView().getBus().setV(390.0);
+        l.getTerminal1().setP(800.0).setQ(400.0); // i = 1324.0969
         assertTrue(l.isOverloaded());
     }
 
@@ -117,38 +117,38 @@ public class CurrentLimitsTest {
     public void testSetterGetter() {
         Line line = createNetwork().getLine("L");
         CurrentLimits currentLimits = line.newCurrentLimits1()
-                                        .setPermanentLimit(100)
+                                        .setPermanentLimit(100.0)
                                             .beginTemporaryLimit()
                                             .setName("20'")
                                             .setAcceptableDuration(20 * 60)
-                                            .setValue(1200)
+                                            .setValue(1200.0)
                                         .endTemporaryLimit()
                                         .beginTemporaryLimit()
                                             .setName("5'")
                                             .setAcceptableDuration(5 * 60)
-                                            .setValue(1400)
+                                            .setValue(1400.0)
                                             .setFictitious(true)
                                         .endTemporaryLimit()
                                         .beginTemporaryLimit()
                                             .setName("1'")
                                             .setAcceptableDuration(60)
-                                            .setValue(1600)
+                                            .setValue(1600.0)
                                         .endTemporaryLimit()
                                     .add();
         try {
-            currentLimits.setPermanentLimit(-0.5f);
+            currentLimits.setPermanentLimit(-0.5);
             fail();
         } catch (ValidationException ignored) {
         }
-        currentLimits.setPermanentLimit(1000f);
-        assertEquals(1000f, currentLimits.getPermanentLimit(), 0.0f);
+        currentLimits.setPermanentLimit(1000.0);
+        assertEquals(1000.0, currentLimits.getPermanentLimit(), 0.0);
         assertEquals(3, currentLimits.getTemporaryLimits().size());
-        assertTrue(Float.isNaN(currentLimits.getTemporaryLimitValue(2)));
+        assertTrue(Double.isNaN(currentLimits.getTemporaryLimitValue(2)));
 
         CurrentLimits.TemporaryLimit temporaryLimit300 = currentLimits.getTemporaryLimit(300);
         assertEquals("5'", temporaryLimit300.getName());
         assertTrue(temporaryLimit300.isFictitious());
-        assertEquals(1400f, temporaryLimit300.getValue(), 0.0f);
+        assertEquals(1400.0, temporaryLimit300.getValue(), 0.0);
         assertEquals(300, temporaryLimit300.getAcceptableDuration());
     }
 }
